Cleanup and re-enable zoom tests
This re-enables the ZoomTestSuite to execute tests for the parts of the
zooming behavior in workbench pages that is actually implemented. This
covers the toggleZoom functionality, which allows to maximize part
stacks. However, some behavior has never been implemented at all or not
in the way the tests currently describe it. This particularly includes
different activation behavior depending on the zoom state when new views
are opened. Thus, those tests are removed, as the current behavior has
been accepted for a long time.

In summary, this changes does:
- enable ZoomTestSuite
- restore unstacked view in zoom tests with the progress view (was
removed when navigator view was removed)
- remove tests related to fast view
- remove tests not fitting to actual and well-established behavior
