+++ tests/org.eclipse.ui.tests/Eclipse UI Tests/org/eclipse/ui/tests/zoom/OpenEditorTest.java
@@ -31,23 +31,6 @@ public class OpenEditorTest extends ZoomTestCase {
 	}
 
 	/**
-	 * <p>Test: Open a new editor while a view is zoomed. Do not force activation.</p>
-	 * <p>Expected result: the page remains zoomed, the view is active</p>
-	 *
-	 * <p>Note: the expected result changed intentionally on 05/04/18</p>
-	 */
-	@Test
-	public void testOpenNewEditorWhileViewZoomed() {
-		close(editor1);
-
-		zoom(stackedView1);
-		openEditor(file1, false);
-
-		assertZoomed(stackedView1);
-		assertActive(stackedView1);
-	}
-
-	/**
 	 * <p>Test: Zoom an editor then open a new editor in the same stack. Do not force activation.</p>
 	 * <p>Expected result: the new editor is zoomed and active</p>
 	 */
@@ -59,47 +42,6 @@ public class OpenEditorTest extends ZoomTestCase {
 		openEditor(file2, false);
 		Assert.assertTrue(isZoomed(editor2));
 		Assert.assertTrue(page.getActivePart() == editor2);
-	}
-
-	/**
-	 * <p>Test: Open an existing editor while a view is zoomed. Do not force activation.</p>
-	 * <p>Expected result: the page remains zoomed, the view is active</p>
-	 */
-	@Test
-	public void testOpenExistingEditorWhileViewZoomed() {
-		zoom(stackedView1);
-		openEditor(file1, false);
-
-		assertZoomed(stackedView1);
-		assertActive(stackedView1);
-	}
-
-	/**
-	 * <p>Test: Open an existing editor while a view is zoomed. Use the activate-on-open mode.</p>
-	 * <p>Expected result: the page is unzoomed, the view is active</p>
-	 */
-	@Test
-	public void testOpenAndActivateExistingEditorWhileViewZoomed() {
-		zoom(stackedView1);
-		openEditor(file1, true);
-
-		assertZoomed(null);
-		assertActive(editor1);
-	}
-
-	/**
-	 * <p>Test: Open a new editor while a view is zoomed. Use the activate-on-open mode.</p>
-	 * <p>Expected result: the page is unzoomed, the view is active</p>
-	 */
-	@Test
-	public void testOpenAndActivateNewEditorWhileViewZoomed() {
-		close(editor1);
-
-		zoom(stackedView1);
-		openEditor(file1, true);
-
-		assertZoomed(null);
-		assertActive(editor1);
 	}
 
 	/**
@@ -127,32 +69,6 @@ public class OpenEditorTest extends ZoomTestCase {
 		openEditor(file2, true);
 
 		assertZoomed(editor2);
-		assertActive(editor2);
-	}
-
-	/**
-	 * <p>Test: Zoom an editor then open an existing editor in a different stack. Do not force activation.</p>
-	 * <p>Expected result: the page remains zoomed and the original editor is active</p>
-	 */
-	@Test
-	public void testOpenExistingEditorInOtherStack() {
-		zoom(editor3);
-		openEditor(file2, false);
-
-		assertZoomed(editor3);
-		assertActive(editor3);
-	}
-
-	/**
-	 * <p>Test: Zoom an editor then open an existing editor in a different stack. Use the activate-on-open mode.</p>
-	 * <p>Expected result: the page is unzoomed and the new editor and active</p>
-	 */
-	@Test
-	public void testOpenAndActivateExistingEditorInOtherStack() {
-		zoom(editor3);
-		openEditor(file2, true);
-
-		assertZoomed(null);
 		assertActive(editor2);
 	}
