+++ tests/org.eclipse.ui.tests/Eclipse UI Tests/org/eclipse/ui/tests/zoom/ZoomPerspectiveFactory.java
@@ -28,6 +28,7 @@ public class ZoomPerspectiveFactory implements IPerspectiveFactory {
 	public static final String STACK1_PLACEHOLDER1 = IPageLayout.ID_PROP_SHEET;
 	public static final String STACK1_VIEW3 = IPageLayout.ID_TASK_LIST;
 	public static final String FASTVIEW1 = IPageLayout.ID_BOOKMARKS;
+	public static final String UNSTACKED_VIEW1 = IPageLayout.ID_PROGRESS_VIEW;
 
 	@Override
 	public void createInitialLayout(IPageLayout layout) {
@@ -40,5 +41,6 @@ public class ZoomPerspectiveFactory implements IPerspectiveFactory {
 		folder.addPlaceholder(STACK1_PLACEHOLDER1);
 		folder.addView(STACK1_VIEW3);
 
+		layout.addView(UNSTACKED_VIEW1, IPageLayout.TOP, 0.5f, IPageLayout.ID_EDITOR_AREA);
 	}
 }
