+++ tests/org.eclipse.ui.tests/Eclipse UI Tests/org/eclipse/ui/tests/zoom/ZoomTestSuite.java
@@ -13,14 +13,12 @@
  *******************************************************************************/
 package org.eclipse.ui.tests.zoom;
 
-import org.junit.Ignore;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 /**
  * A test suite to test the zooming behavior of Eclipse.
  */
-@Ignore
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
 	ZoomedViewActivateTest.class,
