+++ tests/org.eclipse.ui.tests/Eclipse UI Tests/org/eclipse/ui/tests/zoom/ZoomedEditorCloseTest.java
@@ -14,7 +14,6 @@
 package org.eclipse.ui.tests.zoom;
 
 import org.eclipse.ui.IWorkbenchPart;
-import org.junit.Ignore;
 import org.junit.Test;
 
 /**
@@ -30,34 +29,6 @@ public class ZoomedEditorCloseTest extends CloseTest {
 	@Override
 	public IWorkbenchPart getStackedPart2() {
 		return editor2;
-	}
-
-	@Override
-	public IWorkbenchPart getUnstackedPart() {
-		return editor3;
-	}
-
-	/**
-	 * <p>Test: Activate a view, then zoom an unstacked editor and close it.</p>
-	 * <p>Expected result: The previously active Editor becomes active and unzoomed</p>
-	 * <p>Note: This ensures that the activation list is used if there is nothing available
-	 *    in the currently zoomed stack. It also ensures that activation never moves from
-	 *    an editor to a view when an editor is closed.</p>
-	 */
-	@Test
-	@Ignore
-	public void testCloseZoomedUnstackedEditorAfterActivatingView() {
-		System.out.println("Bogus test: we don't unsoom in this case");
-//        IWorkbenchPart previousActive = stackedView1;
-//        IWorkbenchPart zoomedPart = editor3;
-//
-//        page.activate(editor1);
-//        page.activate(previousActive);
-//        zoom(zoomedPart);
-//        close(zoomedPart);
-//
-//        assertZoomed(null);
-//        assertActive(editor1);
 	}
 
 	/**
